+++ changelog.md
@@ -63,3 +63,7 @@ Added edit/add possibility for admin; Added spinner
 Added CanDeactivate guard
 Added auth service and Roles enum; Added CanLoad and CanActivate guards to admin module
 Added localstorage support for Cart service
+
+16.03.2021
+Added json server and concurrently
+Changed products service to work with server using Observable pattern